shifted optical routines to optical.py
