--- zdm/scripts/Path/ics_frbs.py
@@ -1,12 +1,0 @@
-# this file defines the ICS FRBs for which we have PATH info
-# in theory, this file could contain further info.
-
-frblist=['FRB20180924B','FRB20181112A','FRB20190102C','FRB20190608B',
-        'FRB20190611B','FRB20190711A','FRB20190714A','FRB20191001A',
-        'FRB20191228A','FRB20200430A','FRB20200906A','FRB20210117A',
-        'FRB20210320C','FRB20210807D','FRB20211127I','FRB20211203C',
-        'FRB20211212A','FRB20220105A','FRB20220501C',
-        'FRB20220610A','FRB20220725A','FRB20220918A',
-        'FRB20221106A','FRB20230526A','FRB20230708A', 
-        'FRB20230731A','FRB20230902A','FRB20231226A','FRB20240201A',
-        'FRB20240210A','FRB20240304A','FRB20240310A']
